VEOSS-1700 Moved invalidVrnDateController out of auth and into controllers
